+++ CHANGES.rst
@@ -13,6 +13,11 @@ Bug Fixes
 API changes
 ^^^^^^^^^^^
 
+- ``photutils.psf``
+
+  - Added ``output_cols`` as a parameter to ``BasicPSFPhotometry``,
+    ``IterativelySubtractedPSFPhotometry`` and ``DAOPhotPSFPhotometry``. [#745]
+
 
 0.7 (2019-08-14)
 ----------------